start trying to improve analysis performance. Use shorts were possible. Offload unnecesary computation to redis (to be done as needed). Offer options to write less/nothing to disk

+++ sbndcode/DaqAnalysis/Noise.hh
@@ -9,27 +9,27 @@
 namespace daqAnalysis {
 class NoiseSample {
 public:
-  NoiseSample(std::vector<PeakFinder::Peak>& peaks, double baseline, unsigned wvfm_size);
-  NoiseSample(std::vector<std::array<unsigned, 2>> ranges, double baseline): _ranges(ranges), _baseline(baseline) {}
+  NoiseSample(std::vector<PeakFinder::Peak>& peaks, int16_t baseline, unsigned wvfm_size);
+  NoiseSample(std::vector<std::array<unsigned, 2>> ranges, int16_t baseline): _ranges(ranges), _baseline(baseline) {}
   // zero initialize
   NoiseSample(): _baseline(0) {}
 
   NoiseSample Intersection(NoiseSample &other) { return DoIntersection(*this, other, _baseline); }
 
-  double RMS(std::vector<double> &wvfm_self) { return CalcRMS(wvfm_self, _ranges, _baseline); } 
+  float RMS(std::vector<int16_t> &wvfm_self) { return CalcRMS(wvfm_self, _ranges, _baseline); } 
 
-  double Covariance(std::vector<double> &wvfm_self, NoiseSample &other, std::vector<double> &wvfm_other);
-  double Correlation(std::vector<double> &wvfm_self, NoiseSample &other, std::vector<double> &wvfm_other);
-  double SumRMS(std::vector<double> &wvfm_self, NoiseSample &other, std::vector<double> &wvfm_other);
-  static double ScaledSumRMS(std::vector<NoiseSample *>& other, std::vector<std::vector<double> *>& wvfm_other);
+  float Covariance(std::vector<int16_t> &wvfm_self, NoiseSample &other, std::vector<int16_t> &wvfm_other);
+  float Correlation(std::vector<int16_t> &wvfm_self, NoiseSample &other, std::vector<int16_t> &wvfm_other);
+  float SumRMS(std::vector<int16_t> &wvfm_self, NoiseSample &other, std::vector<int16_t> &wvfm_other);
+  static float ScaledSumRMS(std::vector<NoiseSample *>& other, std::vector<std::vector<int16_t> *>& wvfm_other);
 
   std::vector<std::array<unsigned, 2>> *Ranges() { return &_ranges; }
 private:
-  static double CalcRMS(std::vector<double> &wvfm_self, std::vector<std::array<unsigned,2>> &ranges, double baseline);
-  static NoiseSample DoIntersection(NoiseSample &me, NoiseSample &other, double baseline=0.);
+  static float CalcRMS(std::vector<int16_t> &wvfm_self, std::vector<std::array<unsigned,2>> &ranges, int16_t baseline);
+  static NoiseSample DoIntersection(NoiseSample &me, NoiseSample &other, int16_t baseline=0.);
 
   std::vector<std::array<unsigned, 2>> _ranges;
-  double _baseline;
+  int16_t _baseline;
 };
 } // namespace daqAnalysis
 #endif
